time: move time list code to its own file
Just moving code, no significative changes
